requirements.txt
scrapy==1.5.1
boto3==1.9.57
python-dotenv==0.9.1
pystache==0.5.4